dl_app/requirements.txt
tensorflow==2.10.0
transformers
fmpsdk
python-dotenv
pandas
typing 
numpy
python-multipart
uvicorn
fastapi